#78 extract AppIcon enum from App
