+++ jdm-core/src/test/java/jdiskmark/AppTest.java
@@ -42,8 +42,8 @@ class AppTest {
      * About dialog (and taskbar/title-bar) to silently display no icon.
      */
     @ParameterizedTest
-    @EnumSource(App.AppIcon.class)
-    void appIcon_load_isNonNull(App.AppIcon icon) {
+    @EnumSource(AppIcon.class)
+    void appIcon_load_isNonNull(AppIcon icon) {
         assertNotNull(icon.load(),
                 "Icon resource not found on classpath: " + java.util.Arrays.toString(icon.resourcePaths));
     }
